code: finished small project learning about states
You can use the 'by' keyword with variables alongside the remember and rememberSaveable composables to save the state of a variable in an application. The 'remember' composable will update the state of a variable in real time. That memory lost, however, if the composables on screen experience recomposition. This is where 'rememberSaveable' comes in. It remembers state even after recomposition, like when the screen is rotated.